style settings page

+++ frontend/app/components/deleteButton/DeleteButton.tsx
@@ -1,22 +1,27 @@
 import React from 'react';
 import { deleteUser } from '@/app/hooks/deleteUser';
 import { useRouter } from 'next/navigation';
+import { Button } from '@material-tailwind/react';
 
 const DeleteButton = () => {
   const router = useRouter();
 
   const handleDeleteUser = async () => {
-    const success = await deleteUser();
-    if (success) {
-      localStorage.clear();
-      router.push('/api/auth/logout');
-    } else {
-      console.error('Failed to delete user account');
+    if (confirm('Are you sure you want to delete all data?')) {
+      const success = await deleteUser();
+      if (success) {
+        localStorage.clear();
+        router.push('/api/auth/logout');
+      } else {
+        console.error('Failed to delete user account');
+      }
     }
   };
 
   return (
-    <button onClick={handleDeleteUser}>Delete My Account</button>
+    <Button color="red" onClick={handleDeleteUser}>
+      Delete My Account
+    </Button>
   );
 };
